Split CSS files and create attribute object
Split the css file for production testing, the styles.css file now
contains just the page css. script.css contains the css for the range
slider to work on import.

Also added fallbacks to all of the css values if they're not defined.

Added an attributes object to the script.js to condense the conditional
style overrides.